+++ Packages/com.quickeye.ui-toolkit-plus/Runtime/Tab.cs
@@ -6,13 +6,15 @@ namespace QuickEye.UIToolkit
 {
     public class Tab : BaseBindable<bool>
     {
-        public const string ClassName = "tab";
-        public const string TextClassName = "tab__text";
+        public const string ClassName = "qe-tab";
+        public const string TextClassName = ClassName + "__text";
+        public const string CheckedClassName = ClassName + "--checked";
+        public const string UncheckedClassName = ClassName + "--unchecked";
 
         public readonly Reorderable Reorderable = new(ClassName) { LockDragToAxis = true };
 
         private readonly Label _textElement;
-        
+
         private VisualElement _tabContent;
 
         public Tab() : this(null) { }
@@ -20,11 +22,11 @@ namespace QuickEye.UIToolkit
         public Tab(string text)
         {
             this.InitResources();
-            EnableInClassList(ClassName, true);
+            AddToClassList(ClassName);
+
             _textElement = new Label(text);
-            _textElement.EnableInClassList(TextClassName, true);
+            _textElement.AddToClassList(TextClassName);
             Add(_textElement);
-            AddToClassList("tab");
 
             RegisterCallback<PointerDownEvent>(PointerDownHandler);
             IsReorderable = false;
@@ -68,8 +70,8 @@ namespace QuickEye.UIToolkit
 
         private void SetActive(bool isActive)
         {
-            EnableInClassList("tab--checked", isActive);
-            EnableInClassList("tab--unchecked", !isActive);
+            EnableInClassList(CheckedClassName, isActive);
+            EnableInClassList(UncheckedClassName, !isActive);
             TabContent?.ToggleDisplayStyle(isActive);
             if (isActive)
                 DeactivateSiblings();
@@ -88,10 +90,9 @@ namespace QuickEye.UIToolkit
 
         public new class UxmlTraits : BaseBindableTraits<bool, UxmlBoolAttributeDescription>
         {
-            private readonly UxmlBoolAttributeDescription _reorderable = new()
+            private readonly UxmlBoolAttributeDescription _isReorderable = new()
             {
-                name = "Reorderable",
-                defaultValue = false
+                name = "is-reorderable"
             };
 
             private readonly UxmlStringAttributeDescription _text = new() { name = "text" };
@@ -106,7 +107,7 @@ namespace QuickEye.UIToolkit
                 base.Init(ve, bag, cc);
                 var tab = (Tab)ve;
                 tab.Text = _text.GetValueFromBag(bag, cc);
-                tab.IsReorderable = _reorderable.GetValueFromBag(bag, cc);
+                tab.IsReorderable = _isReorderable.GetValueFromBag(bag, cc);
             }
         }
     }