+++ Packages/com.quickeye.ui-toolkit-plus/Runtime/TabGroup.cs
@@ -5,7 +5,10 @@ namespace QuickEye.UIToolkit
 {
     public class TabGroup : VisualElement
     {
-        public const string ClassName = "tab-group";
+        public const string ClassName = "qe-tab-group";
+        public const string VerticalModeClassName = ClassName + "--vertical";
+        public const string HorizontalModeClassName = ClassName + "--horizontal";
+        public const string HorizontalScrollerClassName = ClassName + "__horizontal-scroller";
         private Scroller _horizontalScroller;
         private TabGroupMode _mode;
 
@@ -54,8 +57,7 @@ namespace QuickEye.UIToolkit
             _horizontalScroller.direction = SliderDirection.Horizontal;
             _horizontalScroller.lowButton.SetAction(ScrollView.horizontalScroller.ScrollPageUp, 0, 1);
             _horizontalScroller.highButton.SetAction(ScrollView.horizontalScroller.ScrollPageDown, 0, 1);
-            _horizontalScroller.AddToClassList("tab-group__scroller");
-            hierarchy.Add(_horizontalScroller);
+            _horizontalScroller.AddToClassList(HorizontalScrollerClassName);
 
             UpdateHorizontalScrollerVisibility();
         }
@@ -81,12 +83,21 @@ namespace QuickEye.UIToolkit
 
         private void SetLayoutMode(TabGroupMode mode)
         {
-#if UNITY_2021_2_OR_NEWER
+            EnableInClassList(VerticalModeClassName, mode == TabGroupMode.Vertical);
+            EnableInClassList(HorizontalModeClassName, mode == TabGroupMode.Horizontal);
+            ToggleHorizontalScroller(mode == TabGroupMode.Horizontal);
+
             ScrollView.mode = mode == TabGroupMode.Horizontal ? ScrollViewMode.Horizontal : ScrollViewMode.Vertical;
             ScrollView.verticalScrollerVisibility =
                 mode == TabGroupMode.Horizontal ? ScrollerVisibility.Hidden : ScrollerVisibility.Auto;
-#endif
-            ScrollView.style.alignItems = mode == TabGroupMode.Vertical ? Align.FlexStart : Align.FlexEnd;
+        }
+
+        private void ToggleHorizontalScroller(bool enabled)
+        {
+            if (enabled)
+                hierarchy.Add(_horizontalScroller);
+            else
+                _horizontalScroller.RemoveFromHierarchy();
         }
 
         private new class UxmlFactory : UxmlFactory<TabGroup, UxmlTraits> { }
@@ -97,7 +108,7 @@ namespace QuickEye.UIToolkit
 
             public override IEnumerable<UxmlChildElementDescription> uxmlChildElementsDescription
             {
-                get { yield break; }
+                get { yield return new UxmlChildElementDescription(typeof(VisualElement)); }
             }
 
             public override void Init(VisualElement ve, IUxmlAttributes bag, CreationContext cc)